update spring exercise 1

+++ src/main/java/com/codeup/myblog/PostController.java
@@ -17,7 +17,7 @@ public class PostController {
 
     @GetMapping("/posts/{id}")
     @ResponseBody
-    public String post(@PathVariable int id) {
+    public String post(@PathVariable long id) {
         return "individual post" + id;
     }
